+++ addons/pm_qms_core/models/__init__.py
@@ -1,0 +1,2 @@
+from . import pm_qms_process
+from . import pm_qms_control