+++ addons/pm_qms_core/tests/__init__.py
@@ -1,0 +1,1 @@
+from . import test_pm_qms_core